requirements.txt
Flask==2.2.5
pymongo==4.15.3
python-dotenv==1.0.0
pandas==2.2.3
bcrypt==4.0.1
certifi==2025.10.5
